Create Collapsed Card and handled homePage according to figma
