+++ components/collapsed-card/CollapsedCard.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+import Card from "./Card";
+export default function CollapsedCard() {
+  return (
+    <div className="  w-[98%] h-fit flex flex-row flex-wrap gap-3 justify-center items-center">
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+      <Card />
+    </div>
+  );
+}
